Create a new blog post from a template

This function is designed to work as close as possible to distill::create_post() (and draws on a lot of code from that function). The main difference is the compolusary argument path which gives the path of the template to create the post from. All other arguments work the same as distill::create_post() (unless specified in Arguments).

create_post_from_template(
  path,
  title,
  collection = "posts",
  author = "auto",
  slug = "auto",
  date = Sys.Date(),
  date_prefix = date,
  draft = FALSE,
  edit = interactive()
)

Arguments

path
File path to .Rmd template for post

title
Post title. If there is a title in the template, this will override it.

collection
Collection to create the post within (defaults to "posts")

author
Post author. If not provided, it will be automatically drawn from template. If not in template, it will be automatically drawn from previous post.

slug
Post slug (directory name). Automatically computed from title if not provided.

date
Post date (defaults to current date)

date_prefix
Date prefix for post slug (preserves chronological order for posts within the filesystem). Pass NULL for no date prefix.

draft
Mark the post as a draft (don't include it in the article listing).

edit
Open the post in an editor after creating it.

Note

This function must be called from with a working directory that is within a Distill website.

The output for a post must be distill::distill_article. If there is no output key in the provided template, this will be added to the yaml. If there is an output specified in the template yaml, it must be distill::distill_article, otherwise create_post_from_template will throw an error.

Unlike distill::create_post(), create_post_from_template doesn't automatically provide a description key in the yaml.

Examples

if (FALSE) {
library(distilltools)
# .Rmd templates stored in "templates" directory
create_post_from_template(here::here("templates", "post-template.Rmd"))
}